Cast for print statements in piecewise surface parameter_report

+++ include/eli/geom/surface/piecewise.hpp
@@ -1257,10 +1257,10 @@ namespace eli
               // cycle through all segments to get each bounding box to add
               for (it=key.begin(); it!=key.end(); ++it)
               {
-                printf(" seg: %d \t p: %f \t pk %d\n", i, it->first, it->second);
+                printf(" seg: %d \t p: %f \t pk %d\n", i, (float) (it->first), it->second);
                 ++i;
               }
-              printf(" pmax: %f\n", pmax);
+              printf(" pmax: %f\n", (float) pmax);
               printf("End report\n");
             }
